+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased (develop)
 
+- added: Dismissed promo cards added to `NotificationCenterScene`
 - added: `Edge-ucation` section in `WalletDetailsScene`
 - added: Multiple memo support to `SendScene2`
 - added: Add os version to the "OS" line of logs
 - changed: "MAX" button in `FlipInput` restyled to tertiary and moved next to the text input field
 - fixed: Primary button transforming on the animation on `GettingStartedScene`
+- fixed: `NotificationCenterScene` could crash if there was an undismissed new token notification and the associated wallet no longer existed
 - removed: Disable Fantom transaction list
 
 ## 4.27.0 (2025-05-08)